src/requirements.txt
requests
fastapi
uvicorn
python-multipart
nest_asyncio
python-telegram-bot
dotenv